test: restructure codebase
