+++ package.json
@@ -33,7 +33,8 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "lint": "eslint \"src/**/*.{ts,tsx}\" --fix",
-    "format": "prettier --write"
+    "format": "prettier --write",
+    "coverage": "react-scripts test folder/name --coverage --coverageDirectory='coverage'"
   },
   "husky": {
     "hooks": {
