+++ src/__tests__/router.test.tsx
@@ -1,8 +1,8 @@
 import { render, RenderResult, screen } from "@testing-library/react";
-import HomePage from "components/view/HomePage";
+import HomePage from "components/HomePage";
 import App from "App";
 import Login from "components/view/LoginPage";
-import Signup from "components/view/Signup";
+import SignUp from "components/view/SignUp";
 import TopNav from "components/layout/TopNav";
 import ModalContainer from "components/layout/ModalContainer";
 import { MemoryRouter } from "react-router-dom";
@@ -10,9 +10,9 @@ import { Provider } from "react-redux";
 import store from "store/store";
 
 jest.mock("components/layout/TopNav");
-jest.mock("components/view/HomePage");
+jest.mock("components/HomePage");
 jest.mock("components/view/LoginPage");
-jest.mock("components/view/Signup");
+jest.mock("components/view/SignUp");
 jest.mock("components/layout/ModalContainer");
 
 const renderApp = (entry: string): RenderResult => {
@@ -47,7 +47,7 @@ describe("test routing", () => {
   });
 
   test("should render signup page for signup route", () => {
-    (Signup as jest.Mock).mockImplementation(() => <div>Register Page</div>);
+    (SignUp as jest.Mock).mockImplementation(() => <div>Register Page</div>);
     renderApp("/register");
     expect(screen.getByText("Register Page")).toBeInTheDocument();
   });
