+++ src/components/HomePage/__tests__/CategoryList.test.tsx
