+++ src/components/HomePage/__tests__/HomePage.test.tsx
@@ -3,7 +3,7 @@ import { rest } from "msw";
 import { API, ModalKey } from "utils/constants";
 import { ListResponseDTO } from "utils/DTO";
 import { render, RenderResult, screen, waitForElementToBeRemoved } from "@testing-library/react";
-import HomePage from "components/view/HomePage";
+import HomePage from "components/HomePage";
 import { Provider } from "react-redux";
 import configureStore from "redux-mock-store";
 import userEvent from "@testing-library/user-event";
