+++ src/components/common/__tests__/CustomPagination.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import CustomPagination from "components/HomePage/CustomPagination";
+import CustomPagination from "components/common/CustomPagination";
 
 describe("Custom Pagination", () => {
   const handleSelectPage = jest.fn();
