+++ src/components/view/__tests__/SignUpPage.test.tsx
@@ -5,17 +5,17 @@ import { rest } from "msw";
 import store from "store/store";
 import { Provider } from "react-redux";
 import { MemoryRouter } from "react-router-dom";
-import HomePage from "components/view/HomePage";
-import SignupPage from "components/view/Signup";
+import HomePage from "components/HomePage";
 import { API, AuthTestData } from "utils/constants";
+import SignUpPage from "components/view/SignUp";
 
-jest.mock("components/view/HomePage");
+jest.mock("components/HomePage");
 
-const renderSignupInProvider = (): RenderResult => {
+const renderSignUpInProvider = (): RenderResult => {
   (HomePage as jest.Mock).mockImplementation(() => <div>HomePage</div>);
   return render(
     <Provider store={store}>
-      <SignupPage />
+      <SignUpPage />
       <HomePage />
     </Provider>,
     { wrapper: MemoryRouter },
@@ -40,9 +40,9 @@ beforeAll(() => {
 });
 afterAll(() => server.close());
 
-describe("SignupForm", () => {
+describe("SignUpForm", () => {
   test("should display username, password, email field and submit button", () => {
-    renderSignupInProvider();
+    renderSignUpInProvider();
     expect(screen.getByPlaceholderText(/username/i)).toBeInTheDocument();
     expect(screen.getByPlaceholderText(/password/i)).toBeInTheDocument();
     expect(screen.getByPlaceholderText(/email/i)).toBeInTheDocument();
@@ -50,7 +50,7 @@ describe("SignupForm", () => {
   });
 
   test("should be able to submit form after fill out all fields", () => {
-    renderSignupInProvider();
+    renderSignUpInProvider();
     const signupBtn = screen.getByRole("button", { name: /register/i });
     expect(signupBtn).toBeDisabled();
     userEvent.type(screen.getByPlaceholderText(/username/i), AuthTestData.NAME);
@@ -60,7 +60,7 @@ describe("SignupForm", () => {
   });
 
   test("should have correct error messages when fields are invalid", () => {
-    renderSignupInProvider();
+    renderSignUpInProvider();
     userEvent.type(screen.getByPlaceholderText(/email/i), AuthTestData.INVALID_EMAIL);
     userEvent.type(screen.getByPlaceholderText(/password/i), AuthTestData.INVALID_PASSWORD);
     userEvent.type(screen.getByPlaceholderText(/username/i), AuthTestData.NAME);
@@ -73,7 +73,7 @@ describe("SignupForm", () => {
   });
 
   test("should have correct error message for existed email account", async () => {
-    renderSignupInProvider();
+    renderSignUpInProvider();
     const signupBtn = screen.getByRole("button", { name: /register/i });
     userEvent.type(screen.getByPlaceholderText(/email/i), AuthTestData.EXISTED_EMAIL);
     userEvent.type(screen.getByPlaceholderText(/password/i), AuthTestData.PASSWORD);
@@ -86,7 +86,7 @@ describe("SignupForm", () => {
   });
 
   test("redirect to home page when signup successfully.", async () => {
-    renderSignupInProvider();
+    renderSignUpInProvider();
     const signupBtn = screen.getByRole("button", { name: /register/i });
     userEvent.type(screen.getByPlaceholderText(/email/i), AuthTestData.EMAIL);
     userEvent.type(screen.getByPlaceholderText(/password/i), AuthTestData.PASSWORD);
